requirements.txt
amqp==5.3.1
asgiref==3.7.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
boto3==1.34.11
botocore==1.34.11
CacheControl==0.14.3
cachetools==5.5.2
celery==5.4.0
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.9.1
cron-descriptor==1.4.5
cryptography==45.0.4
daphne==4.2.0
Django==5.0.2
django-celery-beat==2.6.0
django-channels==0.7.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-rest-passwordreset==1.4.0
djangorestframework-simplejwt
django-simple-history==3.7.0
django-storages==1.14.2
psycopg2-binary
python-dotenv
pytz==2024.1
PyYAML==6.0.2
razorpay==1.4.2
redis==5.0.4
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.10.4
service-identity==24.2.0
six==1.17.0
sqlparse==0.4.4
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.9.0
drf-spectacular==0.27.2
Pillow
redis==5.0.4
firebase_admin==6.5.0
django-extensions==3.2.3
pydotplus
pytest
pytest-django
pytest-cov 
pylibmc
python-decouple
pango
WeasyPrint 
jinja2 
matplotlib 
httpx
twilio
razorpay
pypdf
qdrant-client>=1.8.0
python-dateutil

# git+https://<TOKEN>@github.com/anupsawant/piloocore.git